// src/app/model/appstat.ts
export class appstat{
    javaVendor: String;
    javaVersion: String;
    pid: Number;
    uptime: Number;
    heapInit: Number;
    heapMax: Number;
    heapUsed: Number;
    heapCommited:Number;
    nonHeapUsed:Number;
    nonHeapInit: Number;
    nonHeapCommited: Number;
    threadCount: Number;
    daemons:Number;
    
}